🚚 OpenExtensionBubble became LogoBubble
