+++ libs/design-system/src/molecules/LogoBubble/LogoBubble.tsx
@@ -1,10 +1,10 @@
 import { ImageLogo } from '../..';
 
-export type OpenExtensionBubbleProps = {
+export type LogoBubbleProps = {
   onClick?: () => void;
 };
 
-export const OpenExtensionBubble = ({ onClick }: OpenExtensionBubbleProps) => {
+export const LogoBubble = ({ onClick }: LogoBubbleProps) => {
   return (
     <div onClick={onClick} className="plasmo-m-5">
       <ImageLogo className="plasmo-opacity-80 hover:plasmo-opacity-100 hover:plasmo-scale-125 plasmo-transition-all plasmo-transition-200" />
