+++ libs/design-system/src/molecules/LogoBubble/index.ts
@@ -1,0 +1,1 @@
+export * from './LogoBubble';
