--- libs/design-system/src/molecules/openExtensionBubble/index.ts
@@ -1,1 +1,0 @@
-export * from './openExtensionBubble';
